% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmosis_path.R
\name{osmosis_path}
\alias{osmosis_path}
\title{Get Osmosis path}
\usage{
osmosis_path(version = "0.48.3", force = FALSE, quiet = TRUE)
}
\arguments{
\item{version}{A string. The version of Osmosis whose path should be
returned. Defaults to \code{"0.48.3"}, the current latest version. Please check
\href{https://github.com/openstreetmap/osmosis/releases/}{openstreetmap/osmosis releases} for the full
set of available versions.}

\item{force}{A logical. Passed to \code{\link[=download_osmosis]{download_osmosis()}}, whether to overwrite
previously downloaded and cached Osmosis. Defaults to \code{FALSE}.}

\item{quiet}{A logical. Passed to \code{\link[=download_osmosis]{download_osmosis()}}, whether to hide
informative messages or not. Defaults to \code{TRUE}.}
}
\value{
A string, the path to Osmosis.
}
\description{
Returns the path to previously downloaded and cached Osmosis. If it has not
been downloaded yet, downloads it using \code{\link[=download_osmosis]{download_osmosis()}}.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
osmosis_path()
\dontshow{\}) # examplesIf}
}
