% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_osmosis.R
\name{run_osmosis}
\alias{run_osmosis}
\title{Run Osmosis}
\usage{
run_osmosis(osmosis_path, command, echo = TRUE, spinner = TRUE)
}
\arguments{
\item{osmosis_path}{A string. The path to Osmosis.}

\item{command}{A string. The command to run.}

\item{echo}{A logical. Whether to print the standard output and error to the
screen. Defaults to \code{TRUE}.}

\item{spinner}{A logical. Whether to show a reassuring spinner while the
process is running. Defaults to \code{TRUE}.}
}
\value{
Invisibly returns a list containing the exit status of the Osmosis
process, the standard output of the command, the standard error of the
command and whether the process was killed due to a timeout.
}
\description{
Runs Osmosis, given the path to the application and the commands that should
be sent to the command-line tool.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cur_osm <- system.file("extdata/cur.osm.pbf", package = "rosmosis")

fs::file_size(cur_osm)

# cropping the pbf using a bounding box

output_path <- tempfile("cropped_cur", fileext = ".osm.pbf")

osmosis_command <- paste0(
  "--read-pbf ", cur_osm, " ",
  "--bounding-box ",
  "top=-25.4290 left=-49.2792 bottom=-25.4394 right=-49.2629 ",
  "completeWays=yes ",
  "--write-pbf ", output_path
)

run_osmosis(osmosis_path(), osmosis_command, echo = FALSE, spinner = FALSE)
fs::file_size(output_path)
\dontshow{\}) # examplesIf}
}
