% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_osmosis.R
\name{download_osmosis}
\alias{download_osmosis}
\title{Download Osmosis}
\usage{
download_osmosis(version = "0.48.3", quiet = TRUE)
}
\arguments{
\item{version}{A string. Osmosis version.}

\item{quiet}{A logical. Whether to display informative messages or not
(defaults to \code{FALSE}).}
}
\value{
Invisibly returns the path to the downloaded Osmosis file.
}
\description{
Download Osmosis.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
download_osmosis()
\dontshow{\}) # examplesIf}
}
